% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bound}
\alias{bound}
\title{bound}
\format{
A data frame with 18488 rows and 3 variables:
\describe{
  \item{\code{id1}}{integer planning unit ID}
  \item{\code{id2}}{integer the planning unit ID of a planning unit which neighbours the one in id1}
  \item{\code{amount}}{double the relative importance of including the planning unit in id2 if the one in id1 is selected. }
}
}
\usage{
bound
}
\description{
The boundary length file (e.g. bound.dat)
}
\details{
R script to create the data is in [data-raw/bound.R](https://github.com/remi-daigle/marxanr/blob/main/data-raw/bound.R)
}
\keyword{datasets}
