% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{createMarxanFolder}
\alias{createMarxanFolder}
\title{Create a Marxan folder}
\usage{
createMarxanFolder(
  wd = getwd(),
  scen = "demo",
  input = "input",
  output = "output",
  dialog = TRUE
)
}
\arguments{
\item{wd}{The directory in which you would like to create the new Marxan folder. Defaults to your current working directory (i.e. `getwd()`)}

\item{scen}{The name of the new Marxan folder. Defaults to "demo"}

\item{input}{The name of the input folder inside the new Marxan folder. Defaults to "input"}

\item{output}{The name of the input folder inside the new Marxan folder. Defaults to "output"}

\item{dialog}{Overide the system dialog. THe default TRUE will leave the dialog on, and FALSE will turn the dialog off.}
}
\value{
new directory
}
\description{
Create a Marxan folder
}
\examples{
\dontrun{
createMarxanFolder(scen="testingnewfolders")
}
}
