% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marxan.R
\name{downloadMarxan}
\alias{downloadMarxan}
\title{downloadMarxan}
\usage{
downloadMarxan(os = "win", path = getwd(), ver = "4.0.6")
}
\arguments{
\item{os}{character for the operating system. Either "win" (default), "linux", "macos", or "macosM1"}

\item{path}{character string for the destination path. Defaults to the current working directory.}

\item{ver}{character string for the version number. Default is "4.0.6"}
}
\value{
files
}
\description{
Download Marxan from the releases on the [github repository](https://github.com/Marxan-source-code/marxan/releases)
}
\examples{
downloadMarxan(os="win")
}
