% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{land}
\format{
A data frame with 3 rows and 2 variables:
\describe{
  \item{\code{name}}{character Country name}
  \item{\code{geometry}}{list the simple features column (sfc) describing the land}
}
}
\usage{
land
}
\description{
An sf object describing the land near the MPA workflow vignette. Cropped from the rnaturalearth package.
}
\details{
R script to create the data is in [data-raw/land.R](https://github.com/remi-daigle/marxanr/blob/main/data-raw/land.R)
}
\keyword{datasets}
