% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\docType{class}
\name{marxanParams-class}
\alias{marxanParams-class}
\title{marxanParams: class that represents the Marxan Input Paramaters}
\value{
class
}
\description{
marxanParams: class that represents the Marxan Input Paramaters
}
\section{Slots}{

\describe{
\item{\code{BLM}}{numeric.}

\item{\code{PROP}}{numeric.}

\item{\code{RANDSEED}}{numeric.}

\item{\code{NUMREPS}}{integer.}

\item{\code{NUMITNS}}{integer.}

\item{\code{STARTTEMP}}{numeric.}

\item{\code{NUMTEMP}}{numeric.}

\item{\code{COSTTHRESH}}{numeric.}

\item{\code{THRESHPEN1}}{numeric.}

\item{\code{THRESHPEN2}}{numeric.}

\item{\code{INPUTDIR}}{character.}

\item{\code{PUNAME}}{character.}

\item{\code{SPECNAME}}{character.}

\item{\code{PUVSPRNAME}}{character.}

\item{\code{BOUNDNAME}}{character.}

\item{\code{SCENNAME}}{character.}

\item{\code{SAVERUN}}{integer.}

\item{\code{SAVEBEST}}{integer.}

\item{\code{SAVESUMMARY}}{integer.}

\item{\code{SAVESCEN}}{integer.}

\item{\code{SAVETARGMET}}{integer.}

\item{\code{SAVESUMSOLN}}{integer.}

\item{\code{SAVEPENALTY}}{integer.}

\item{\code{SAVELOG}}{integer.}

\item{\code{OUTPUTDIR}}{character.}

\item{\code{RUNMODE}}{integer.}

\item{\code{MISSLEVEL}}{integer.}

\item{\code{ITIMPTYPE}}{integer.}

\item{\code{HEURTYPE}}{integer.}

\item{\code{CLUMPTYPE}}{integer.}

\item{\code{VERBOSITY}}{integer.}

\item{\code{SAVESOLUTIONSMATRIX}}{integer.}
}}

