% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pu}
\alias{pu}
\title{pu}
\format{
A data frame with 2794 rows and 2 variables:
\describe{
  \item{\code{id}}{character planning unit id}
  \item{\code{geometry}}{list the simple features column (sfc) describing the planning units}
}
}
\usage{
pu
}
\description{
hexagonal planning units created for marxanr package demonstrations
}
\details{
R script to create the data is in [data-raw/pu.R](https://github.com/remi-daigle/marxanr/blob/main/data-raw/pu.R)
}
\keyword{datasets}
