% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pu_dat}
\alias{pu_dat}
\title{pu_dat}
\format{
A data frame with 2794 rows and 5 variables:
\describe{
  \item{\code{id}}{character the planning unit ID}
  \item{\code{cost}}{double the cost to protect the planning unit}
  \item{\code{status}}{double the planning unit status}
  \item{\code{xloc}}{double longitude}
  \item{\code{yloc}}{double latitude}
}
}
\usage{
pu_dat
}
\description{
The marxan formatted planning unit file (e.g. pu.dat)
}
\details{
R script to create the data is in [data-raw/pu_dat.R](https://github.com/remi-daigle/marxanr/blob/main/data-raw/pu_dat.R)
}
\keyword{datasets}
