% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec}
\alias{spec}
\title{spec}
\format{
A data frame with 78 rows and 4 variables:
\describe{
  \item{\code{id}}{integer the species or conservation feature ID}
  \item{\code{prop}}{double the proportion to be protected}
  \item{\code{spf}}{double the species penalty factor}
  \item{\code{name}}{character the name of the species or conservation feature}
}
}
\usage{
spec
}
\description{
The marxan formatted species file (e.g. spec.dat). Data was generated using ecological data layers from [Serdynska et al. 2021](https://open.canada.ca/data/en/dataset/2d9cce9a-d634-4b49-879f-87c40c52acf2) will be used in this example. These data are the available bioregional-scale ecological and human use data used to inform draft Marine Protected Area (MPA) design for the Scotian Shelf-Bay of Fundy Bioregion. Please see the [original technical report](https://publications.gc.ca/collections/collection_2021/mpo-dfo/Fs97-6-3382-eng.pdf) for further details. All spatial data has been modified/simplified for demonstration purposes and should not be used for any other purpose.
}
\details{
R script to create the data is in [data-raw/spec_and_puvspr.R](https://github.com/remi-daigle/marxanr/blob/main/data-raw/spec_and_puvspr.R)
}
\keyword{datasets}
