% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{writeParams}
\alias{writeParams}
\title{Write Marxan Input Parameter File (input.dat) to disk}
\usage{
writeParams(inpudat = marxanr::newParams(), file = "input.dat")
}
\arguments{
\item{inpudat}{a `marxanParams` class object to be written to disk. A `marxanParams` class object can be generated using `newParams()`}

\item{file}{a character string naming the file}
}
\description{
Write Marxan Input Parameter File (input.dat) to disk
}
\examples{
inputdat <- newParams(BLM = 1000)
writeParams(inputdat,"input.dat")
}
