% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{newParams}
\alias{newParams}
\title{Create new Marxan Input Parameter File (input.dat)}
\usage{
newParams(
  BLM = 1,
  PROP = 0.5,
  RANDSEED = -1,
  NUMREPS = 10L,
  NUMITNS = 1000000L,
  STARTTEMP = -1,
  NUMTEMP = 10000,
  COSTTHRESH = 0,
  THRESHPEN1 = 14,
  THRESHPEN2 = 1,
  INPUTDIR = "input",
  PUNAME = "pu.dat",
  SPECNAME = "spec.dat",
  PUVSPRNAME = "puvspr.dat",
  BOUNDNAME = "bound.dat",
  SCENNAME = "output",
  SAVERUN = 3L,
  SAVEBEST = 3L,
  SAVESUMMARY = 3L,
  SAVESCEN = 3L,
  SAVETARGMET = 3L,
  SAVESUMSOLN = 3L,
  SAVEPENALTY = 3L,
  SAVELOG = 2L,
  OUTPUTDIR = "output",
  RUNMODE = 1L,
  MISSLEVEL = 1L,
  ITIMPTYPE = 0L,
  HEURTYPE = -1L,
  CLUMPTYPE = 0L,
  VERBOSITY = 3L,
  SAVESOLUTIONSMATRIX = 3L
)
}
\arguments{
\item{SAVESOLUTIONSMATRIX}{}
}
\value{

}
\description{
Create new Marxan Input Parameter File (input.dat)
}
\examples{
# Accepting all default values
inputdat <- newParams()

# Accepting all default values except BLM
inputdat <- newParams(BLM = 1000)
}
