% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pu}
\alias{pu}
\title{pu}
\format{
A data frame with 2794 rows and 2 variables:
\describe{
  \item{\code{id}}{character planning unit ID}
  \item{\code{geometry}}{list the simple features column (sfc) describing the planning units}
}
}
\usage{
pu
}
\description{
hexagonal planning units created for marxanr package demonstrations
}
\details{
DETAILS
}
\keyword{datasets}
