% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputfiles.R
\name{writeParams}
\alias{writeParams}
\title{Write Marxan Input Parameter File (input.dat) to disk}
\usage{
writeParams(inpudat, file)
}
\arguments{
\item{inpudat}{}
}
\value{

}
\description{
Write Marxan Input Parameter File (input.dat) to disk
}
\examples{
inputdat <- newParams(BLM = 1000)
writeParams(inputdat,"input.dat")
}
